refactor(nyz): add new compatibility file in ding top level
