+++ ding/torch_utils/network/nn_module.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.init import xavier_normal_, kaiming_normal_, orthogonal_
 from typing import Union, Tuple, List, Callable
-from ding import torch_gt_131
+from ding.compatibility import torch_gt_131
 
 from .normalization import build_normalization
